% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{example_log}
\alias{example_log}
\alias{patients}
\title{Patients eventlog}
\format{Eventlog containing 500 patient cases}
\usage{
patients
}
\description{
Artifical eventlog about patients
}
\keyword{datasets}
