% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{hospital}
\alias{hospital}
\title{Hospital log}
\format{Eventlog containing 1143 cases}
\source{
<https://data.4tu.nl/repository/uuid:d9769f3d-0ab0-4fb8-803b-0d1120ffcf54>
}
\usage{
hospital
}
\description{
Real life log of a Dutch academic hospital, originally intended for use in the first Business Process Intelligence Contest (BPIC 2011)
}
\keyword{datasets}
