% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{hospital_billing}
\alias{hospital_billing}
\title{Hospital billing log}
\format{Eventlog containing 100000 cases}
\source{
<https://data.4tu.nl/repository/uuid:76c46b83-c930-4798-a1c9-4be94dfeb741>
}
\usage{
hospital_billing
}
\description{
The 'Hospital Billing' event log was obtained from the financial modules of the ERP system of a regional hospital. The event log contains events that are related to the billing of medical services that have been provided by the hospital. Each trace of the event log records the activities executed to bill a package of medical services that were bundled together. The event log does not contain information about the actual medical services provided by the hospital.
The 100,000 traces in the event log are a random sample of process instances that were recorded over three years. Several attributes such as the 'state' of the process, the 'caseType', the underlying 'diagnosis' etc. are included in the event log. Events and attribute values have been anonymized. The time stamps of events have been randomized for this purpose, but the time between events within a trace has not been altered.
More information about the event log can be found in the related publications.
}
\references{
Mannhardt F., de Leoni M., Reijers H.A., van der Aalst W.M.P. (2017) Data-Driven Process Discovery - Revealing Conditional Infrequent Behavior from Event Logs. In: Dubois E., Pohl K. (eds) Advanced Information Systems Engineering. CAiSE 2017. Lecture Notes in Computer Science, vol 10253. Springer, Cham
}
\keyword{datasets}
