% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{sepsis}
\alias{sepsis}
\title{Sepsis Cases - Event Log}
\format{An object of class \code{eventlog} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 15214 rows and 35 columns.}
\source{
<http://data.4tu.nl/repository/uuid:915d2bfb-7e84-49ad-a286-dc35f063a460>
}
\usage{
sepsis
}
\description{
This real-life event log contains events of sepsis cases from a hospital. Sepsis is a life threatening condition typically caused by an infection. One case represents the pathway through the hospital. The events were recorded by the ERP (Enterprise Resource Planning) system of the hospital. There are about 1000 cases with in total 15,000 events that were recorded for 16 different activities. Moreover, 39 data attributes are recorded, e.g., the group responsible for the activity, the results of tests and information from checklists. Events and attribute values have been anonymized. The time stamps of events have been randomized, but the time between events within a trace has not been altered.
}
\references{
Analyzing the Trajectories of Patients with Sepsis using Process Mining (Felix Mannhardt and Daan Blinde, CEUR Workshop Proceedings Vol-1859 RADAR+EMISA 2017)
}
\keyword{datasets}
