% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{traffic_fines}
\alias{traffic_fines}
\title{Road Traffic Fina Management Process Log}
\format{Eventlog containing 150370 cases}
\source{
<https://data.4tu.nl/repository/uuid:270fd440-1057-4fb9-89a9-b699b47990f5>
}
\usage{
traffic_fines
}
\description{
TReal-life event log of an information system managing road traffic fines.
}
\references{
Mannhardt, F., de Leoni, M., Reijers, H.A. et al. Computing (2016) 98: 407. https://doi.org/10.1007/s00607-015-0441-1
}
\keyword{datasets}
