% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{example_log}
\alias{example_log}
\alias{example_log}
\alias{patients}
\title{Example eventlog}
\format{Eventlog containing 12766 events}
\usage{
example_log

patients
}
\description{
Artifical eventlog

Artifical eventlog about patients
}
\keyword{datasets}
